Agent Twin/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
google-generativeai==0.3.0
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
pydub==0.25.1
django-cors-headers==4.3.1

# Voice cloning with ElevenLabs
elevenlabs==0.2.27

# Optional: For better audio transcription
# openai-whisper==20231117
# speech-recognition==3.10.0

